Move generators to a separate package
